Added Omnisharp
I have added the use of omnisharp to handle c# code. It requires to add the omnisharp server
on the location specified on init-csharp.el At some point it will be taken out into a
system variable. Also, the non-stable version of melpa had to be added because of a dependency
of omnisharp on the dash package.