Starting modifications due to windows
Now that I will have emacs on windows as well a few modifications
need to be taken into account. I hace created two separate inits for
windows and *nix systems. Also, because of investigations, now I have
changed the way that the fullscreen is toggled and how the line
numbers are displayed. Finally added the powershell mode